feat: move global styling to HOC
